A little bit of ssize_t just to change things up, and also to work again.

+++ src/pyjion/absint.cpp
@@ -1302,7 +1302,7 @@ void AbstractInterpreter::branchRaise(const char *reason, size_t curByte) {
         m_comp->emit_trace_exception();
 
     // number of stack entries we need to clear...
-    auto count = m_stack.size() - entryStack.size();
+    auto count = static_cast<ssize_t>(m_stack.size() - entryStack.size());
 
     auto cur = m_stack.rbegin();
     for (; cur != m_stack.rend() && count >= 0; cur++) {
